structuralize provider code
